# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Anteater Adventure Club",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/admin/events": {
      "get": {
        "summary": "Admin Events",
        "operationId": "admin_events_api_admin_events_get",
        "parameters": [
          {
            "name": "quarter_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Quarter Id"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Search"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_EventPrivate_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Event",
        "operationId": "create_event_api_admin_events_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventWrite-Input"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventPrivate"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/series": {
      "post": {
        "summary": "Create Series",
        "operationId": "create_series_api_admin_series_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeriesWrite-Input"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/members": {
      "post": {
        "summary": "Add Member",
        "operationId": "add_member_api_admin_members_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberPrivate"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Member List",
        "operationId": "member_list_api_admin_members_get",
        "parameters": [
          {
            "name": "quarter_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Quarter Id"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Search"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "all",
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/officers": {
      "get": {
        "summary": "Officers",
        "operationId": "officers_api_admin_officers_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Items_OfficerView_"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Grant Officer",
        "operationId": "grant_officer_api_admin_officers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OfficerGrant"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/quarters": {
      "get": {
        "summary": "Admin Quarters",
        "operationId": "admin_quarters_api_admin_quarters_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_QuarterPrivate_"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Quarter",
        "operationId": "create_quarter_api_admin_quarters_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuarterWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuarterPrivate"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/media": {
      "post": {
        "summary": "Upload",
        "operationId": "upload_api_admin_media_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_api_admin_media_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/board/terms": {
      "get": {
        "summary": "Admin Terms",
        "operationId": "admin_terms_api_admin_board_terms_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Items_BoardTermView_"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add Term",
        "operationId": "add_term_api_admin_board_terms_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TermWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardTermView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EventWrite-Input": {
        "properties": {
          "expected_revision": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Revision"
          },
          "quarter_id": {
            "type": "integer",
            "title": "Quarter Id"
          },
          "name": {
            "type": "string",
            "maxLength": 150,
            "minLength": 1,
            "title": "Name"
          },
          "kind": {
            "type": "string",
            "enum": [
              "regular",
              "meeting",
              "picnic",
              "retreat"
            ],
            "title": "Kind",
            "default": "regular"
          },
          "destination": {
            "type": "string",
            "maxLength": 200,
            "title": "Destination",
            "default": ""
          },
          "description": {
            "type": "string",
            "maxLength": 10000,
            "title": "Description",
            "default": ""
          },
          "starts_at": {
            "type": "string",
            "format": "date-time",
            "title": "Starts At"
          },
          "ends_at": {
            "type": "string",
            "format": "date-time",
            "title": "Ends At"
          },
          "signups_enabled": {
            "type": "boolean",
            "title": "Signups Enabled",
            "default": true
          },
          "opens_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opens At"
          },
          "closes_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closes At"
          },
          "arrival_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Arrival At"
          },
          "departure_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Departure At"
          },
          "return_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Return At"
          },
          "packing": {
            "items": {
              "type": "string",
              "maxLength": 200
            },
            "type": "array",
            "maxItems": 50,
            "title": "Packing",
            "default": []
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/Question"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Questions",
            "default": []
          },
          "photo_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Id"
          },
          "miles": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,10}|(?=[\\d.]{1,13}0*$)\\d{0,10}\\.\\d{0,2}0*$)"
              }
            ],
            "title": "Miles",
            "default": "0"
          },
          "gas_price": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d{0,3}0*$"
              }
            ],
            "title": "Gas Price",
            "default": "0"
          },
          "rate_override": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d{0,6}0*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rate Override"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "quarter_id",
          "name",
          "starts_at",
          "ends_at"
        ],
        "title": "EventWrite"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MemberInput": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 254
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "phone": {
            "type": "string",
            "maxLength": 40,
            "title": "Phone",
            "default": ""
          },
          "student": {
            "type": "boolean",
            "title": "Student",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "MemberInput"
      },
...
      "Page_EventPrivate_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EventPrivate"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "Page[EventPrivate]"
      },
...
      "SeriesView": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "definition": {
            "$ref": "#/components/schemas/SeriesWrite-Output"
          },
          "occurrences": {
            "items": {
              "$ref": "#/components/schemas/SeriesOccurrence"
            },
            "type": "array",
            "title": "Occurrences"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "revision",
          "definition",
          "occurrences"
        ],
        "title": "SeriesView"
      },
      "SeriesWrite-Input": {
        "properties": {
          "expected_revision": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Revision"
          },
          "event": {
            "$ref": "#/components/schemas/EventWrite-Input"
          },
          "kind": {
            "type": "string",
            "enum": [
              "weekly",
              "alternating"
            ],
            "title": "Kind",
            "default": "weekly"
          },
          "starts_on": {
            "type": "string",
            "format": "date",
            "title": "Starts On"
          },
          "until": {
            "type": "string",
            "format": "date",
            "title": "Until"
          },
          "start_time": {
            "type": "string",
            "format": "time",
            "title": "Start Time"
          },
          "end_time": {
            "type": "string",
            "format": "time",
            "title": "End Time"
          },
          "weekdays_a": {
            "items": {
              "type": "integer",
              "maximum": 6.0,
              "minimum": 0.0
            },
            "type": "array",
            "maxItems": 7,
            "minItems": 1,
            "title": "Weekdays A"
          },
          "weekdays_b": {
            "items": {
              "type": "integer",
              "maximum": 6.0,
              "minimum": 0.0
            },
            "type": "array",
            "maxItems": 7,
            "title": "Weekdays B",
            "default": []
          },
          "excluded": {
            "items": {
              "type": "string",
              "format": "date"
            },
            "type": "array",
            "maxItems": 366,
            "title": "Excluded",
            "default": []
          },
          "effective_from": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective From"
          },
          "request_id": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Request Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "event",
          "starts_on",
          "until",
          "start_time",
          "end_time",
          "weekdays_a",
          "request_id"
        ],
        "title": "SeriesWrite"
      },
...
    }
  }
}